Alterada a autenticação, mudanças em algumas rotas
Administradores agora não possuem modelo próprio, há uma verificação na própria autenticação para ver se um usuário é um administrador.

+++ server/routes/turnos.js
@@ -1,25 +1,33 @@
 const express = require('express');
 const router = express.Router();
 const _ = require('lodash');
-var { Turno } = require('../models/turno');
-var { authenticateAdmin } = require('../middleware/authenticateAdmin');
+let {
+    Turno
+} = require('../models/turno');
+let {
+    authenticate
+} = require('../middleware/authenticate');
 
-router.post('/register', authenticateAdmin, (req, res) => {
-    var turno = _.pick(req.body, ['turnos']).turnos;
-    Turno.create(turno).then((turno) => {
-        res.status(200).send(turno);
-    }, (e) => {
-        res.status(400).send(e);
-    })
+router.post('/register', authenticate, (req, res) => {
+    if (req.isAdmin) {
+        let turno = _.pick(req.body, ['turnos']).turnos;
+        Turno.create(turno).then((turno) => {
+            res.status(200).send(turno);
+        }, (e) => {
+            res.status(400).send(e);
+        })
+    }
 })
 
-router.delete('/:id', authenticateAdmin, (req, res) => {
-    var id = req.params.id;
-    Turno.findByIdAndRemove(id).then((turno) => {
-        res.status(200).send(turno);
-    }, (e) => {
-        res.status(400).send(e);
-    })
+router.delete('/:id', authenticate, (req, res) => {
+    if (req.isAdmin) {
+        let id = req.params.id;
+        Turno.findByIdAndRemove(id).then((turno) => {
+            res.status(200).send(turno);
+        }, (e) => {
+            res.status(400).send(e);
+        })
+    }
 })
 
 module.exports = router;
